model 클래스 분리
