+++ src/main/kotlin/blackjack/model/Rank.kt
@@ -1,0 +1,17 @@
+package blackjack.model
+
+enum class Rank(val value: Int) {
+    TWO(2),
+    THREE(3),
+    FOUR(4),
+    FIVE(5),
+    SIX(6),
+    SEVEN(7),
+    EIGHT(8),
+    NINE(9),
+    TEN(10),
+    JACK(10),
+    QUEEN(10),
+    KING(10),
+    ACE(1), // 에이스는 1 또는 11로 계산 필요
+}
