+++ src/main/kotlin/blackjack/model/Suit.kt
@@ -1,0 +1,8 @@
+package blackjack.model
+
+enum class Suit(val description: String) {
+    SPADES("스페이드"),
+    HEARTS("하트"),
+    DIAMONDS("다이아몬드"),
+    CLUBS("클로버"),
+}
