+++ src/main/kotlin/blackjack/view/ResultView.kt
@@ -1,14 +1,18 @@
 package blackjack.view
 
 class ResultView {
-    fun renderPlayerInitOutput(names: String) =
-        render(names + INIT_CARDS)
+    fun renderPlayerInitOutput(names: String) = render(names + INIT_CARDS)
 
-    fun renderPlayerCardsOutput(name: String, cards: String) =
-        render("${name}카드 : ${cards}")
+    fun renderPlayerCardsOutput(
+        name: String,
+        cards: String,
+    ) = render("${name}카드 : $cards")
 
-    fun renderPlayerCardsResultOutput(name: String, cards: String, sum: Int) =
-        render("${name}카드 : $cards - 결과 : $sum")
+    fun renderPlayerCardsResultOutput(
+        name: String,
+        cards: String,
+        sum: Int,
+    ) = render("${name}카드 : $cards - 결과 : $sum")
 
     private fun render(message: String) {
         println(message)
